ACE test release: beta 1
Use "python ace.py --help" for options or install to
src/ace/ of your asss build environment to use with
the asss makefile.

See the ACE wiki for documentation and examples.

Major functionality:
- Generate full asss modules (especially the entry point function)
from a builder language that lets the user specify specific asss patterns.
- Wrap asss constructs including callbacks, interfaces, commands,
and advisers by automatically registering and unregistering them
in the correct order.
- Specify what interfaces your module needs to successfully load,
and ACE will fully handle and log any failures to obtain the interfaces.
- Automatic handling of "dynamic" per-arena-data and per-player-data,
allowing full structures to only be allocated when they apply.
- Automatically prototype functions.